put output and print fcns to a different src and h. Also small cleanup to controls
