+++ include/parameters.h
@@ -34,4 +34,7 @@
 // Control parameters
 #define CTRL_PID_OMEGA_RPS  8.0f
 #define CTRL_PID_DAMPING    8.0f
+#define CTRL_PID_KP         POINT_MASS_KG * CTRL_PID_OMEGA_RPS*CTRL_PID_OMEGA_RPS/ENV_GRAVITY_MPS2
+#define CTRL_PID_KD         2.0 * CTRL_PID_OMEGA_RPS * CTRL_PID_DAMPING/ENV_GRAVITY_MPS2
+#define CTRL_PID_KI         CTRL_PID_KD
 #endif // PARAMETERS_H_
